feat(web): add support for includeNullValues for other range components
