+++ packages/web/src/utils/index.js
@@ -103,3 +103,24 @@ export const getNullValuesQuery = fieldName => ({
 		},
 	},
 });
+
+export const getRangeQueryWithNullValues = (value, props) => {
+	let query = null;
+	const rangeQuery = {
+		range: {
+			[props.dataField]: {
+				gte: value[0],
+				lte: value[1],
+				boost: 2.0,
+			},
+		},
+	};
+	if (props.includeNullValues) {
+		query = {
+			bool: {
+				should: [rangeQuery, getNullValuesQuery(props.dataField)],
+			},
+		};
+	} else query = rangeQuery;
+	return query;
+};
